fix(brick-kit): update user_id of analytics when logout and login
Closes EASY_HUB-337
