+++ packages/brick-kit/src/auth.spec.ts
@@ -1,7 +1,13 @@
+import { userAnalytics } from "@next-core/easyops-analytics";
 import { authenticate, getAuth, logout } from "./auth";
+import { initAnalytics } from "./core/initAnalytics";
 import { resetPermissionPreChecks } from "./internal/checkPermissions";
 
+jest.spyOn(userAnalytics, "initialized", "get").mockReturnValue(true);
+jest.mock("./core/initAnalytics");
 jest.mock("./internal/checkPermissions");
+
+const mockInitAnalytics = initAnalytics as jest.Mock;
 
 describe("auth", () => {
   it("should work", () => {
@@ -18,10 +24,12 @@ describe("auth", () => {
       userInstanceId: "abc",
       accessRule: "cmdb",
     });
+    expect(mockInitAnalytics).toBeCalled();
 
     expect(resetPermissionPreChecks).not.toBeCalled();
     logout();
     expect(getAuth()).toEqual({});
     expect(resetPermissionPreChecks).toBeCalled();
+    expect(mockInitAnalytics).toBeCalled();
   });
 });
