+++ packages/brick-kit/src/auth.ts
@@ -1,4 +1,6 @@
 import { AuthInfo } from "@next-core/brick-types";
+import { userAnalytics } from "@next-core/easyops-analytics";
+import { initAnalytics } from "./core/initAnalytics";
 import { resetPermissionPreChecks } from "./internal/checkPermissions";
 
 const auth: AuthInfo = {};
@@ -12,6 +14,11 @@ export function authenticate(newAuth: AuthInfo): void {
     loginFrom: newAuth.loginFrom,
     accessRule: newAuth.accessRule,
   });
+
+  // re-init analytics to set user_id
+  if (userAnalytics.initialized) {
+    initAnalytics();
+  }
 }
 
 /**
@@ -32,6 +39,9 @@ export function logout(): void {
   auth.userInstanceId = undefined;
   auth.accessRule = undefined;
   resetPermissionPreChecks();
+
+  // re-init analytics to clear user_id
+  initAnalytics();
 }
 
 /**
